refactor(cli): move show stats implementation with other commands
