+++ ado/cli/resources/data_container/show_stats.py
