+++ ado/cli/resources/sample_store/show_stats.py
